server crud moved to shop-manager. servers view and route removed.
