+++ src/views/ShopsView.vue
@@ -247,8 +247,8 @@ function handleFundsInput(event) {
 			<p class="font-medium">No servers found</p>
 			<p class="text-sm mt-1">
 				You need to create a server first before adding shops.
-				<router-link to="/servers" class="text-blue-600 hover:text-blue-800 underline">
-					Go to Servers
+				<router-link to="/shop-manager" class="text-blue-600 hover:text-blue-800 underline">
+					Go to Shop Manager
 				</router-link>
 			</p>
 		</div>
